+++ NEWS.md
@@ -2,6 +2,21 @@
 
 NEW FEATURES
 
+* Added `awst` and `gene_filter` methods for `SummarizedExperiment` class.
+* Added tests for the new methods.
+
+SIGNIFICANT USER-VISIBLE CHANGES
+
+* Matrix methods of `awst` and `gene_filter` now return a gene by sample matrix.
+
+BUG FIXES
+
+None.
+
+# awst 0.99.0
+
+NEW FEATURES
+
 * Added a `NEWS.md` file to track changes to the package.
 * Added a vignette to illustrate the main usage of the package.